Added boxes and some animations
Added animations for Punching up aswell as the animation that is a reaction to a hard puch.

Also added in boxes, boxes are created when socket emits "new_box" but will be changed in future to when fame reaches max. boxes cannot be opened yet and apart from being drawn onto canvas does nothing at all. Three types of boxes exist in the game